# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/post/create": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "description": "Create New Post with Token.",
                "operationId": "5c3ae80191b021c88c1a131009e4764a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "New Post Title"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Content of the new post"
                                    },
                                    "like": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/post/{id}/like": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Increase likes of a post",
                "description": "Increase likes of a post by post ID.",
                "operationId": "615db20155538da65f2ba4334d2a748b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "like_count": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Likes updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/show/{id}": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user by ID",
                "description": "Display the specified user.",
                "operationId": "d9ccb8b4cff41806cb4d58e895d0b0a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Post": {
                "title": "Post",
                "required": [
                    "user_id",
                    "title",
                    "body"
                ],
                "properties": {
                    "id": {
                        "description": "Post ID",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Post title",
                        "type": "string"
                    },
                    "body": {
                        "description": "Post content",
                        "type": "string"
                    },
                    "like": {
                        "description": "Post Like",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "username",
                    "country_code",
                    "mobile_number",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "username": {
                        "description": "User Name",
                        "type": "string",
                        "example": "mohamad ashrafi"
                    },
                    "country_code": {
                        "description": "User Country Code",
                        "type": "integer",
                        "example": "98"
                    },
                    "mobile_number": {
                        "description": "User Phone Number",
                        "type": "integer",
                        "example": "9102403100"
                    },
                    "password": {
                        "description": "User Password",
                        "type": "integer",
                        "example": "12345678"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token to access these endpoints",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}